GHC.Internal.TH.Ppr

Description

contains a prettyprinter for the Template Haskell datatypes

Documentation

nestDepth :: Int

type Precedence = Int

appPrec :: Precedence

opPrec :: Precedence

unopPrec :: Precedence

funPrec :: Precedence

qualPrec :: Precedence

sigPrec :: Precedence

noPrec :: Precedence

parensIf :: Bool -> Doc -> Doc

pprint :: Ppr a => a -> String

class Ppr a where

Minimal complete definition

ppr

Methods

ppr :: a -> Doc

ppr_list :: [a] -> Doc

ppr_sig :: Name -> Type -> Doc

pprFixity :: Name -> Fixity -> NamespaceSpecifier -> Doc

pprNamespaceSpecifier :: NamespaceSpecifier -> Doc

pprPatSynSig :: Name -> PatSynType -> Doc

Pretty prints a pattern synonym type signature

pprPatSynType :: PatSynType -> Doc

Pretty prints a pattern synonym's type; follows the usual conventions to print a pattern synonym type compactly, yet unambiguously. See the note on PatSynType and the section on pattern synonyms in the GHC user's guide for more information.

pprPrefixOcc :: Name -> Doc

isSymOcc :: Name -> Bool

pprInfixExp :: Exp -> Doc

pprExp :: Precedence -> Exp -> Doc

pprFields :: [(Name, Exp)] -> Doc

pprMaybeExp :: Precedence -> Maybe Exp -> Doc

pprMatchPat :: Pat -> Doc

pprGuarded :: Doc -> (Guard, Exp) -> Doc

pprBody :: Bool -> Body -> Doc

pprClause :: Bool -> Clause -> Doc

pprLit :: Precedence -> Lit -> Doc

bytesToString :: [Word8] -> String

pprString :: String -> Doc

pprPat :: Precedence -> Pat -> Doc

ppr_dec

Arguments

:: Bool
declaration on the toplevel?

-> Dec

-> Doc

ppr_deriv_strategy :: DerivStrategy -> Doc

ppr_overlap :: Overlap -> Doc

ppr_data

Arguments

:: Bool
declaration on the toplevel?

-> Doc

-> Cxt

-> Maybe Name

-> Doc

-> Maybe Kind

-> [Con]

-> [DerivClause]

-> Doc

ppr_newtype

Arguments

:: Bool
declaration on the toplevel?

-> Doc

-> Cxt

-> Maybe Name

-> Doc

-> Maybe Kind

-> Con

-> [DerivClause]

-> Doc

ppr_type_data

Arguments

:: Bool
declaration on the toplevel?

-> Doc

-> Cxt

-> Maybe Name

-> Doc

-> Maybe Kind

-> [Con]

-> [DerivClause]

-> Doc

ppr_typedef :: String -> Bool -> Doc -> Cxt -> Maybe Name -> Doc -> Maybe Kind -> [Con] -> [DerivClause] -> Doc

ppr_deriv_clause :: DerivClause -> Doc

ppr_tySyn :: Doc -> Maybe Name -> Doc -> Type -> Doc

ppr_tf_head :: TypeFamilyHead -> Doc

ppr_bndrs :: PprFlag flag => Maybe [TyVarBndr flag] -> Doc

commaSepApplied :: [Name] -> Doc

pprForall :: [TyVarBndr Specificity] -> Cxt -> Doc

pprForallVis :: [TyVarBndr ()] -> Cxt -> Doc

pprForall' :: PprFlag flag => ForallVisFlag -> [TyVarBndr flag] -> Cxt -> Doc

pprRecFields :: [(Name, Strict, Type)] -> Type -> Doc

pprGadtRHS :: [(Strict, Type)] -> Type -> Doc

pprVarBangType :: VarBangType -> Doc

pprBangType :: BangType -> Doc

pprVarStrictType :: (Name, Strict, Type) -> Doc

Deprecated: As of template-haskell-2.11.0.0, VarStrictType has been replaced by VarBangType. Please use pprVarBangType instead.

pprStrictType :: (Strict, Type) -> Doc

Deprecated: As of template-haskell-2.11.0.0, StrictType has been replaced by BangType. Please use pprBangType instead.

pprType :: Precedence -> Type -> Doc

pprParendType :: Type -> Doc

pprInfixT :: Precedence -> Type -> Doc

pprParendTypeArg :: TypeArg -> Doc

isStarT :: Type -> Bool

pprTyApp :: Precedence -> (Type, [TypeArg]) -> Doc

fromTANormal :: TypeArg -> Maybe Type

pprFunArgType :: Type -> Doc

data ForallVisFlag

Constructors

ForallVis

ForallInvis

data TypeArg

Constructors

TANormal Type

TyArg Kind

split :: Type -> (Type, [TypeArg])

pprTyLit :: TyLit -> Doc

class PprFlag flag where

Methods

pprTyVarBndr :: TyVarBndr flag -> Doc

pprBndrVis :: BndrVis -> Doc -> Doc

pprCtxWith :: Ppr a => (Precedence -> a -> Doc) -> [a] -> Doc

pprCxt :: Cxt -> Doc

ppr_ctx_preds_with :: Ppr a => (Precedence -> a -> Doc) -> Precedence -> [a] -> Doc

ppr_cxt_preds :: Precedence -> Cxt -> Doc

where_clause :: [Dec] -> Doc

showtextl :: Show a => a -> Doc

hashParens :: Doc -> Doc

quoteParens :: Doc -> Doc

sepWith :: Doc -> (a -> Doc) -> [a] -> Doc

commaSep :: Ppr a => [a] -> Doc

commaSepWith :: (a -> Doc) -> [a] -> Doc

semiSep :: Ppr a => [a] -> Doc

semiSepWith :: (a -> Doc) -> [a] -> Doc

unboxedSumBars :: Doc -> SumAlt -> SumArity -> Doc

bar :: Doc